Adding updated code

+++ processing_scripts/ray_cast_bmode_data.py
@@ -12,7 +12,7 @@ def ray_cast_image(image):
 
     for i in range(image.shape[1]):
         for j in j_range:
-            if image[j, i] != 0:
+            if image[j, i] != 12:
                 rays[j, i] = 1
                 break
 
@@ -42,34 +42,44 @@ def generate_raycasted_db(db_path, output_dir):
             os.makedirs(save_dir)
 
         split_path = os.path.join(db_path, split)
+        print(split_path)
 
         us_images = [item for item in os.listdir(split_path) if "label" not in item]
         us_labels = [item for item in os.listdir(split_path) if "label" in item]
 
         for (image_name, label_name) in zip(us_images, us_labels):
 
-            try:
-                labelpath = os.path.join(split_path, label_name)
-                label = np.array(Image.open(labelpath))
-                ray_casted_labels = ray_cast_image(label)
+            suffix = image_name.split("_")[-1]
+            label_name_expected = image_name.replace(suffix, "label_" + suffix)
 
-                # do not save empty images for training
-                if np.sum(ray_casted_labels) == 0:
-                    continue
+            labelpath = os.path.join(split_path, label_name)
 
-                ray_casted_labels = cv2.dilate(ray_casted_labels, np.ones((10, 10)), iterations=1)
+            assert label_name == label_name_expected, "non correct label"
+            assert os.path.exists(labelpath), "Inexisting label path"
 
-                image_save_path = os.path.join(save_dir, image_name)
-                label_save_path = os.path.join(save_dir, label_name)
+            image_save_path = os.path.join(save_dir, image_name)
+            label_save_path = os.path.join(save_dir, label_name)
 
-                copy2(os.path.join(split_path, image_name), image_save_path)
-                Image.fromarray(ray_casted_labels).save(label_save_path)
-            except:
-                print("something wrong with: ", image_name)
+            if os.path.exists(image_save_path) and os.path.exists(label_save_path):
+                continue
+
+            label = np.array(Image.open(labelpath))
+            ray_casted_labels = ray_cast_image(label)
+
+            # do not save empty images for training
+            if np.sum(ray_casted_labels) == 0:
+                continue
+
+            ray_casted_labels = cv2.dilate(ray_casted_labels, np.ones((10, 10)), iterations=1)
+
+
+            copy2(os.path.join(split_path, image_name), image_save_path)
+            Image.fromarray(ray_casted_labels).save(label_save_path)
+
 
 
 generate_raycasted_db(db_path="E:/NAS/jane_project/unet_data/segmentation_network_data/full_labels",
-                       output_dir="E:/NAS/jane_project/unet_data/segmentation_network_data/ray_casted_db")
+                       output_dir="E:/NAS/jane_project/unet_data/segmentation_network_data/ray_casted_labels")
